fix: DB関連のHooksを定義し直した
