--- src/features/DisplayFile/hooks/useFile.ts
@@ -1,14 +1,0 @@
-import { useLiveQuery } from "dexie-react-hooks";
-import { db } from "@/db/db";
-
-function useFile() {
-  const displayFiles = useLiveQuery(async () => {
-    const displayFileUrls = db.displayFiles.toArray();
-
-    return displayFileUrls;
-  });
-
-  return { displayFiles };
-}
-
-export default useFile;
